go/src/samsaradev.io/infra/dataplatform/dagster/projects/datamodel/marimo/notebook_task/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "notebook-task-marimo"
version = "0.1.0"
description = "Marimo notebook for running notebook_task_table compute logic locally"
requires-python = ">=3.9"
dependencies = [
    "marimo>=0.8.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "databricks-sql-connector>=3.0.0",
    "python-dotenv>=1.0.0",
    "pyspark>=3.3.0",
    "databricks-connect>=13.0.0,<15.5.0",
    "setuptools>=80.9.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.6.0",
    "isort>=5.10.1",
    "pytest",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.marimo]
# Marimo configuration
